Fix difficult to pick up egg when standing atop nest at high framerates
The player must be grounded before the pick up animation can be triggered. But, when standing on top of a nest, collision jitter causes the grounded state to be unstable from frame to frame, especially at high framerates. (Nests are one of the two objects in the game with the weird hull collision type, the other being the static boulder.)

The workaround is to check whether the player is grounded as long as the player holds down the pickup key, not just the moment they initiated the press.